+++ scripts/common_cmdline.sh
@@ -68,7 +68,7 @@ function compileFull
     "$SOLC" "${args[@]}" "${files[@]}" >/dev/null 2>"$stderr_path"
     local exit_code=$?
     local errors; errors=$(grep -v -E \
-        -e 'Warning: This is a pre-release compiler version|Warning: Experimental features are turned on|pragma experimental ABIEncoderV2|^ +--> |^ +\||^[0-9]+ +\| ' \
+        -e 'Warning: This is a pre-release compiler version|Warning: Experimental features are turned on|^ +--> |^ +\||^[0-9]+ +\| ' \
         -e '^No text representation found.$' < "$stderr_path"
     )