+++ test/libsolidity/semanticTests/functionTypes/store_function.sol
@@ -29,5 +29,5 @@ contract C {
 // gas irOptimized code: 19000
 // gas legacy: 79492
 // gas legacy code: 69600
-// gas legacyOptimized: 77587
+// gas legacyOptimized: 77611
 // gas legacyOptimized code: 28600