+++ test/libsolidity/syntaxTests/array/length/builtin_abi_namespace_call.sol
@@ -1,0 +1,5 @@
+contract C {
+    uint[uint(abi())] x;
+}
+// ----
+// TypeError 5462: (22-33): Invalid array length, expected integer literal or constant expression.