+++ test/libsolidity/syntaxTests/constantEvaluator/non_function_magic_variable_call.sol
@@ -1,0 +1,6 @@
+contract C {
+    uint constant L = uint(abi("")[0]);
+    uint[L] x;
+}
+// ----
+// TypeError 5462: (62-63): Invalid array length, expected integer literal or constant expression.