+++ test/libsolidity/syntaxTests/pragma/experimental_abiencoderv2_removed.sol
@@ -1,0 +1,3 @@
+pragma experimental ABIEncoderV2;
+// ----
+// SyntaxError 2044: (0-33): The experimental ABIEncoderV2 pragma has been removed from the compiler.